Move the MetricsDropdown component to its own file.
